Moved the heatmap function into a separate file, added the libraries and the pycache folders to the .gitignore
