+++ .gitignore
@@ -1,1 +1,3 @@
 **/.ipynb_checkpoints/
+forward-solvers/__pycache__/
+libraries/
