+++ forward-solvers/phd-of-rho-and-z-compute-fluence.py
@@ -13,7 +13,6 @@ file = '../libraries/Vts.dll'
 clr.AddReference(os.path.abspath(file))
 import numpy as np
 import plotly.graph_objects as go
-import plotly.express as px
 from Vts import *
 from Vts.Common import *
 from Vts.Extensions import *
@@ -28,9 +27,8 @@ from Vts.MonteCarlo.Detectors import *
 from Vts.MonteCarlo.Factories import *
 from Vts.MonteCarlo.PhotonData import *
 from Vts.MonteCarlo.PostProcessing import *
-from System import Array, Double, Object, Func, Math
-clr.AddReference("System.Core")
-from System.Linq import Enumerable
+from System import Array, Double, Math
+from graph_tools import heatmap
 
 solver = DistributedPointSourceSDAForwardSolver()
 
@@ -82,26 +80,6 @@ log_phd = [Math.Log(f) for f in phdOfRhoAndZ]
 size = len(zs)
 # split into rows
 phdRowsToPlot = np.array([log_phd[i:i+size] for i in range(0, len(log_phd), size)])
-
-# Heatmap function to convert the data into a heat map
-def heatmap(values, x, y, x_label="", y_label="", title=""):
-    """Create a heatmap chart."""
-    # values should be a 2D array-like (list of lists or 2D numpy array)
-    fig = go.Figure(data=go.Heatmap(
-        z=values,
-        x=x,
-        y=y,
-        transpose=True,
-        colorscale='Hot',
-        colorbar=dict(title=title)
-    ))
-    fig.update_layout(
-        title=title,
-        xaxis_title=x_label,
-        yaxis_title=y_label,
-        yaxis_autorange='reversed'
-    )
-    return fig
 
 fluenceChart = heatmap(phdRowsToPlot.tolist(), allRhos.tolist(), list(zs), "ρ [mm]", "z [mm]", "log(phd(ρ, z) [mm-2])")
 fluenceChart.show(renderer="browser")
